Chore: Add Annotate

See merge request itsmycargo/imc-react-api!1033

Former-commit-id: 4f719646b5ba011e730314b5c33cce7ccb5b5740

+++ app/models/conversation.rb
@@ -4,7 +4,7 @@ class Conversation < ApplicationRecord
   belongs_to :shipment
   belongs_to :user
   belongs_to :tenant
-  belongs_to :manager, class_name: "User", optional: true
+  belongs_to :manager, class_name: 'User', optional: true
   has_many :messages
   def update_unreads
     unread = messages.where(read: false).count
@@ -18,11 +18,23 @@ class Conversation < ApplicationRecord
   end
 
   def self.clear_old_conversations
-    self.all.each do |convo|
-      if convo.shipment.nil?
-        convo.destroy
-      end
+    all.each do |convo|
+      convo.destroy if convo.shipment.nil?
     end
   end
-
 end
+
+# == Schema Information
+#
+# Table name: conversations
+#
+#  id           :bigint(8)        not null, primary key
+#  shipment_id  :integer
+#  tenant_id    :integer
+#  user_id      :integer
+#  manager_id   :integer
+#  created_at   :datetime         not null
+#  updated_at   :datetime         not null
+#  last_updated :datetime
+#  unreads      :integer
+#
